Added component: Accept Payment which presents the customer with a status page upon successfully completing the payment process. This component uses the following files: - /api/classes/accept.php - /webroot/accept.php - /webroot/templates/default/xhtml/accept.xsl Additionally the component relies on the RetrieveImage component for resizing mPoint's logo to fit the screensize of the customer's mobile device.
Started new module: Receipt. This module offers the customer a receipt for the transaction.

Added component: E-Mail Receipt which provides business logic for sending the customer a receipt via E-Mail as part of the Receipt Module.
This component uses the following files:
- /api/classes/accept.php
- /api/classes/email_receipt.php
- /webroot/email.php
- /webroot/sys/send_mail.php
- /webroot/templates/default/xhtml/email.xsl
Additionally the component relies on the RetrieveImage component for resizing mPoint's logo to fit the screensize of the customer's mobile device.


git-svn-id: svn://svn.cellpointmobile.com/mpoint/server/trunk@11 86c063f6-31ef-4d33-b9ba-d2d3e4924f03