Added a new part: Mobile Web to mPoint's Buy API. This API provides a simple and convenient means to allow Mobile Internet Sites and Mobile Applications with access to mPoint's Payment flow. The new module: MobileWeb will validate the client input and return the "Select Credit Card" page from the Payment module to the customer. This module consist of the following files: - /api/classes/mobile_web.php - /webroot/buy/web.php - /webroot/_test/web.php
The introduction of this new API part has meant that the inheritance hierarchy for the other parts of the Buy API: SMS and Call Centre has been changed.
The Call Centre Module now extends the SMS Purchase Module which in turn extends the Mobile Web Module.

Updated the global include file: inc/global.php to start the user session and initialise the Output module from the PHP5 API for the Mobile Web Module.

git-svn-id: svn://svn.cellpointmobile.com/mpoint/server/trunk@25 86c063f6-31ef-4d33-b9ba-d2d3e4924f03