Added default stylesheet for mobile phones.
Fixed bug in the Image Retrieval module which prevented the mPoint logo from being returned.

Fixed bug in component: E-Mail from module: Receipt so that the customer's input is stored in a session and the input field is automatically repopulated if an error is found in the e-mail address.

Optimised all XSL Stylesheets after testing on the following mobiles phones:
- Nokia 6230
- Nokia E60
- Sony Ericsson K750i
- Sony Ericsson K790i
mPoint is useful on all of these phones, however on the Nokia 6230 the layout is somewhat "weird" as it doesn't support different width for table columns. The payment pages are still usable though.

git-svn-id: svn://svn.cellpointmobile.com/mpoint/server/trunk@26 86c063f6-31ef-4d33-b9ba-d2d3e4924f03