Added information text to the product overview.
Added new database field: decimals to table: Country_Tbl.
This field holds information about how many decimals that a price should be displayed with in the given country.
Current settings are 0 for Denmark and 2 for USA.
The new field can be accessed via the method: getDecimals().
Additionally this addition has meant updates to the the static produce method for the following classes:
- ClientConfig
- TxnInfo

Changed mPoint CSS classes to start the 2nd word with a capital letter:
- mPoint_Label
- mPoint_Status
- mPoint_Info
- mPoint_Number

Added new method for authorising a credit card transaction via DIBS using the "Custom Pages" functionality.
The pages in the dibs folder needs to be uploaded to DIBS' server along with all credit card images, the client logo and the "Payment succeeded" Icon

git-svn-id: svn://svn.cellpointmobile.com/mpoint/server/trunk@30 86c063f6-31ef-4d33-b9ba-d2d3e4924f03