Added new control variable: Mode to the Client Configuration. This variable controls how the Payment Transaction is processed: 0 = Production 1 = Test Mode with pre-filled card Info 2 = Certification Mode, same as test mode but card info is not pre-filled The mode that a transaction is processed in is also stored in the Transaction Log.
Added new control field: position INT4 to database table: System.Card_Tbl.
The field identifies the order that the cards are displayed in, the smaller the position is the closer to the top it is.

Fixed bugs caused by the use of colspan attributes on the Nokia 60E and N95.
These phones can only handle one colspan attribute for each row, if multiple td tags defines a colspan attribute the phones are unable to render the table properly.

git-svn-id: svn://svn.cellpointmobile.com/mpoint/server/trunk@43 86c063f6-31ef-4d33-b9ba-d2d3e4924f03