Fixed bug which caused disabled products to be displayed on the Order Overview and the Product Selection pages.
Added updated mPoint - Overview documentation which describes the different mPoint APIs

Added documentation specific for implementing mPoint with Mobile Internet Sites.
This documentation is a cut down version of the mPoint - Overview that is mainly intended at possible competitors.

Added graphics for the mPoint frontpage that should be used for all mPoint documents

git-svn-id: svn://svn.cellpointmobile.com/mpoint/server/trunk@45 86c063f6-31ef-4d33-b9ba-d2d3e4924f03