Fixed bugs in the implementation of Premium SMS as a Payment Method. The Cellpoint Mobile Payment component will now generate a Callback to mPoint which in turn will send out an e-mail and / or SMS receipt.
Fixed bug when restoring the data object holding the Transaction Information from the database which prevented the e-mail address retrieved from the database from being stored in the object.
This in turn meant that the E-Mail Receipt wasn't send out as no recipient was defined.

Fixed bug in method: constHeaders from class: EMailReceipt which caused the e-mail receipt to be sent twice.

git-svn-id: svn://svn.cellpointmobile.com/mpoint/server/trunk@49 86c063f6-31ef-4d33-b9ba-d2d3e4924f03