Fixed bug in method: send from class: Callback which sent the wrong HTTP Header for the HTTP POST when sending the callback data.
Fixed bug in the SMS component of the Buy API which prevented GoMobile ID'et for the incoming MO-SMS from being logged.
This meant that SMS wouldn't be available as a Payment Option later in the flow when selecting the Card.

Added the customer's mobile number to the error message if the initial automatic address lookup fails.
This clarifies to the customer why the page would load with the initial error message.

Re-Created the text translation files from the code to ensure that all messages can be found.

Expanded on the notes for manual tasks when uploading mPoint in the Ant Build file: build.xml

git-svn-id: svn://svn.cellpointmobile.com/mpoint/server/trunk@53 86c063f6-31ef-4d33-b9ba-d2d3e4924f03