Changed the design of the "Products" page in the mPoint - Mobile Shop so the XHTML listing the products are simpler. Additionally products are now purchased using an "Add to Basket" link rather than having the customer enter the Quantity for each product.
Fixed issue with the SMS Receipt sent out by the Callback Module so the text is no longer UTF8 Decoded twice.

Fixed bug in the DIBS component from the Callback Module if the "language" parameter wasn't returned by DIBS.
The component now uses the "lang" parameter if "language" isn't returned and default to British English if none of the parameters are available.

git-svn-id: svn://svn.cellpointmobile.com/mpoint/server/trunk@57 86c063f6-31ef-4d33-b9ba-d2d3e4924f03