Modified the SQL script for creating a new client to ensure all mPoint settings are properly defined
Added example of the DIBS files that must be uploaded to the Client's DIBS Account for mPoint to work.

Added example of PayPal Mobile for use in the certification process in the US.

Updated the test component with new logos.

Added example of the HTML Form used to start a payment transaction with DIBS

Added logo files

git-svn-id: svn://svn.cellpointmobile.com/mpoint/server/trunk@59 86c063f6-31ef-4d33-b9ba-d2d3e4924f03