Extended the Web and Call Centre APIs so a client may provide the customer's e-mail address. This has resulted in the "recipient" parameter being renamed to "mobile" to prevent confusion.
Renamed the parameter: "recipient" to "mobile" in the Callback API to reflect the new name in the Web and Call Centre APIs.

Made the E-Mail receipt optional and moved the "Send E-Mail Receipt" page to the end of the payment flow after the payment has been successfully completed.

Improved the XSL files used to render the payment pages to improve the customization options.
All pages now contains one or more section ids which can be used to define the look for specific parts of each page.
For instance, the input form on the "Delivery Information" page has been wrapped in a div tag with id="delivery" to allow the background colour for the input fields to be changed.
Additionally all listings now uses one of the following CSS classes:
- mPoint_Even (for even rows)
- mPoint_Uneven (for uneven rows)

Updated the API documents to reflect the new changes.

Fixed bug in the global include file so the underlying Output Renderer is called with the default value: screen for the media attribute in the link tag.

git-svn-id: svn://svn.cellpointmobile.com/mpoint/server/trunk@61 86c063f6-31ef-4d33-b9ba-d2d3e4924f03