Fixed type error for constant: iEMAIL_REJECTED_STATE
Removed UTF-8 decoding of the message body from method: sendEmbeddedLink in class: SurePay.

Updated the documentation to include description of the "orderid" parameter and language code "sv" for Swedish.
Added Swedish translation files and prepared the following DIBS files for Swedish translation:
- payment.html
- accept.html
- decline.html

Improved the error logging in the DIBS component from the Callback module.

Switched the SQL scripts for the Client and Country setup so the proper SQL queries are in each file.

git-svn-id: svn://svn.cellpointmobile.com/mpoint/server/trunk@68 86c063f6-31ef-4d33-b9ba-d2d3e4924f03