Begun development of version 1.30 which intends to add the following new features: - E-Money based Stored Value Account - Stored Credit Card associated with the End-User's Account (requires approval from Acquirer, i.e. PBS, Euroline, Teller etc.) - Web based End-User administration interface allowing the user to:    - View Transaction History    - Manage Account    - Top-Up E-Money    - Transfer E-Money to another user (existing or new) - Create account through one of the following methods:    - Create account via Web (requires validation code via SMS to ensure ownership of MSISDN)    - Create account via Web or Mobile based on an E-Money transfer    - Create account while paying for goods by checking "Save Card Info" on the payment page - Allow payment via Web by running payment flow in a popup - Compatibility with ARTS Stored Value Account specifications
==============================================================================================================================

Added End-User Account which allows a user to top-up the account with e-money and store one or more credit cards with the account.
This greatly simplifies the payment experience for returning customers as payments can be authorised simply by selecting the prepaid account or a stored credit card and entering a predefine password.
Account creation has been incorporated into the standard payment flow by providing the end-user with a "Store Card Info" option on the payment page.
Checking this box will automatically create a new end-user account in mPoint and request the user to enter a password for the account.
To support these features the a new schema: EndUser has been added to the database containing the following tables:
- Account_Tbl, data table for storing the information associated with the End-User account
- Transaction_Tbl, data table for keeping track of all E-Money based transactions
- Card_Tbl, data table for storing all card information. The table doesn't actually store the card number but only card meta-data: masked card no, expiry date and ticket associated with the card
- CLAccess_Tbl, link table identifying which Clients the End-User account may be used through. If no entries are found in this table the Account is available for all clients in the country.
Implementing the End-User Account has also changed the class inheritance hierarchy as a number of classes now relies on class: Home in order to handle common operations related to the End-User's account.

Added preliminary AJAX enabled web based administration interface for the end-user. The admin interface currently allows the end-user to view the following:
- Transaction History:
    - E-Money Transfers
    - E-Money Top-Ups
    - E-Money Purchases
    - Credit Card Purchases
- Account Information
- Stored Credit Cards
- View Account Balance
To simplify login, the admin interface automatically attempts to determine the end-user's country based on the IP address. In order to support this behaviour a new table: IPRange_Tbl has been added to database schema: System.

Translated DIBS payment pages into Swedish

Added DIBS documentation

Added configuration file: php.ini for auto generating documentation based on the PHP Doc comments in the code

Renamed global XSL Stylesheet: header.xsl to mobile.xsl to reflect that it is only used when accessing mPoint via the mobile phone.

Simplified the logic of the image component so that it expects the width/heigh of an image to be provided as part of the image name.
Image URLs must now be in the format: {WIDTH}x{HEIGHT}_{TYPE}_{PHP SESSION ID}.png.
Additionally images are returned in PNG format rather than JPEG format by default.

Improved the intelligence of the "Save Delivery" component so that the end-user is automatically redirected to "checkout" (which in turn redirects to "order overview") if only one shipping method is available for the client. This allows the user to essentially skip a step thereby providing an improved user experience.

git-svn-id: svn://svn.cellpointmobile.com/mpoint/server/trunk@69 86c063f6-31ef-4d33-b9ba-d2d3e4924f03