Fixed possibility for a bug in the SQL query of method: getAccountID from class: EndUserAccount by adding aliases to all fields referenced in the query.
Fixed bug when outputting the SQL query of method: getCountryFromIP from class: General for debug purposes.

Clarified the SQL query in method: getAccountID from class: Home by removing use of unnecessary aliases.

Changed the way Transfer Codes are generated in method: _constNewAccountMessage from class: Transfer

Fixed bug in purchase API: TopUp (/buy/topmenu.php) which caused the End-User Account ID to be lost when the transaction as re-created for the system client.

Added login validation checks to the following files:
- home/sys/make_transfer.php
- home/sys/save_info.php
- home/sys/save_mobile.php
- home/sys/save_password.php
- home/sys/send_code.php
- home/sys/send_link.php

Performed minor code clean-up of the login component so the XML / XSL files are more in line with the rest of mPoint.
The login component has also been specified as its own sub-package in the PHP Doc comments for each of the following files:
- login/content.php
- login/default.php
- login/sys/auth.php
- login/topmenu.php

Changed all paths in the main AJAX XSL file: ajax/default.xsl to reference images, stylesheets etc. using absolute paths.
This ensures that the file may be used from other paths than /

Removed the spacing under the main headline (h1) to allow progress steps to be listed just beneath the page headline.
In order to preserve whitespace between the headline and the page content, a line break has been added to the following files:
- ajax/home/activate.xsl
- ajax/home/history.xsl
- ajax/home/my_account.xsl
- ajax/home/topup.xsl
- ajax/home/transfer.xsl

Made minor clarifications to the comments in the following classes:
- MyAccount

git-svn-id: svn://svn.cellpointmobile.com/mpoint/server/trunk@76 86c063f6-31ef-4d33-b9ba-d2d3e4924f03