Added association between a stored card and the client for which it was stored to allow an End-User to store the same card to be stored for multiple clients with different tickets. When a card is stored a link is now also automatically created between the End-User Account and the Client for which the card is stored if no such link already exists.
Added complete ATOMICITY to the transfer controller: /home/sys/make_transfer.php.
This ensures that the database is rolled back unless all of the following operations succeed:
1. Create new user account
2. Send transfer notification
3. Transfer money
The first 2 steps are only applicable if the recipient doesn't have an account yet.

Renamed the method constructing SMTP Headers: constHeaders from class: Home to: constSMTPHeaders to prevent confusion with the other constHeaders method which constructed HTTP Headers.
This change has affected the following methods:
- EMailReceipt::sendReceipt
- MyAccount::sendLink
- SurePay::notifyClient
- Transfer::sendNewAccountEMail

Renamed the method constructing HTTP Headers: constHeaders from class: General to: constHTTPHeaders to prevent confusion with the other constHeaders method which constructed SMTP Headers.
This change has affected the following methods:
- Callback::performCallback
- CellpointMobile::initCallback
- Delivery::lookupViaInterflora
- DIBS::authTicket
- DIBS::capture
- DIBS::delTicket
- DIBS::initCallback

Fixed bug when dynamically constructing the SQL query in method: _genActivationCode from class: MyAccount which caused the constructed query to be invalid.

Fixed bug in the SQL query of method: activateCode from class: MyAccount which caused the query to be invalid

Fixed bug in method: valChecksum from class: MyAccount which caused validation using an E-Mail address to fail.

Fixed bug in the SMS Purchase API which caused an error to be raised when constructing the Data Object holding the Transaction Information due to an invalid static call to method: getAccountID from class: EndUserAccount.

Fixed bug in the Pay via Account component from the Cellpoint Mobile module which caused authentication to always fail due to the wrong value being passed to method: auth from class: General.
Also fixed a minor bug during validation which could raise an E_NOTICE due to array key: prepaid not always existing in the auto global $_POST array.

Fixed bug in the back-end component for making a transfer which caused the wrong status code to be returned when validating the recipient as an e-mail address.

Added support for automatically recaching the transaction history when making changes to My Account.
This has affected the following back-end components:
- /home/sys/save_info.php
- /home/sys/save_mobile.php

Fixed bugs in the htaccess file for serving images which caused an HTTP Code 404 to be returned rather than the image due to an error in the regular expression after the image urls have changed to include the width / height that the image should be resized to.

Fixed bug which prevented the Web Interface from working in Internet Explorer.
This has meant changing the conditions for instantiating the Output Control Object in: inc/include.php to specifically only be instantiated for pages displayed on the mobile phone.
Additionally the content-type returned been changed from application/xhtml+xml to text/html for the following files:
- /templates/default/xhtml/mobile.xsl
- /templates/default/xhtml/pay/card.xsl
- /templates/default/xhtml/cpm/payment.xsl
Finally the Output Control Object is no longer instantiated manually on the following pages:
- /new/default.php
- /home/accept.php

Fixed bug in /internal/unauthorized.php which didn't clear the Client side cache when unauthorized access to the system was detected.

Fixed bug in the viewer for the "My Account" page: /templates/default/ajax/home/my_account.xsl which caused the displayed data to disappear when changing folder.

Fixed bug in the viewer for the Top Menu which caused a whitespace to be missing when displaying the balance on the End-User's Prepaid Account.

Fixed bug which caused an error: "Address already registered to a user account" when a Transfer Code hasn't yet been entered during the process for creating a new account.
Methods: valMobile and valEMail from class: MyAccount now returns a different code when the provided Address is already registered to an account that hasn't yet been activated (an account where no password has been defined) which allows the controller: /new/sys/create_account.php to take this scenario into account when determining what status code / message to return.

git-svn-id: svn://svn.cellpointmobile.com/mpoint/server/trunk@84 86c063f6-31ef-4d33-b9ba-d2d3e4924f03