Fixed bug in method: generatePage of the AJAX Client Library: client.js which didn't properly account for Internet Explorer stripping JavaScript code when the transformed HTML was inserted into the page.
Changed the image size for the card logos on the Transaction History page so they match the logos displayed for the Stored Cards on the My Account page.

Fixed bug in /img/get_image.php which caused the Session ID to be set to a wrong value if the Session ID wasn't part of the filename for the image that should be displayed.

Fixed minor glitch in the row layout for "Purchase History" on the Transaction History page due to card logos being displayed for each row.

Fixed bug in the viewer for the My Account page which caused all stored cards to have the same row colour.

git-svn-id: svn://svn.cellpointmobile.com/mpoint/server/trunk@86 86c063f6-31ef-4d33-b9ba-d2d3e4924f03