Fixed bug in the viewers of "My Account" and "History" which prevented these pages from working in Mozilla Firefox on Mac. The bug resulted in the following error being logged in the debug window: "WARNING - content - Error: Component returned failure code: 0x80004003 (NS_ERROR_INVALID_POINTER) [nsIDOMNSHTMLElement.innerHTML] at line: 1016"
Removed the "Order Number" from all accept pages and receipts if no Order Number was provided by the Client when invoking mPoint.
This has affected the following viewers:
- dibs/accept.html
- dibs/accept_multi.html
- templates/default/xhtml/pay/accept.xsl
Additionally the following methods have been changed:
- Callback::sendSMSReceipt
- EMailReceipt::sendReceipt

Merged the 2 login controllers: send_otp.php and auth.php into one file which accepts the following parameters:
- countryid
- username
- password
- otp
A One Time Password will be sent via SMS to the user's mobile phone if "otp" and a popup shown is undefined where as the user will be logged in if "otp" is provided.
This provides a more clean API for logging in rather than requiring the client to first call send_otp.php and then auth.php.

git-svn-id: svn://svn.cellpointmobile.com/mpoint/server/trunk@90 86c063f6-31ef-4d33-b9ba-d2d3e4924f03