Changed the code to use: foreach ($aErrCd as $tag => $code) rather than: while (list($tag, $code) = each($aErrCd) ) to ensure that the array of status codes are always automatically reset when looping. This has affected the following controller files: - home/sys/make_transfer.php - home/sys/save_info.php - home/sys/save_mobile.php - home/sys/save_password.php - home/sys/send_code.php - home/sys/send_link.php - new/sys/create_account.php - new/sys/save_mobile.php - login/sys/auth.php
Implemented support for requiring Password and Confirmation Codes to authenticate a transfer. A Confirmation Code will automatically be sent to the sender's mobile number if this has been defined in the profile. The code must then be entered in conjunction with the password in a popup window to authenticate the transfer.

Implemented initial support for fees have been added, fees may be defined for the following actions:
- Top-Up
- Transfer
- Withdrawal
For Transfers, fees may be defined for for each country based on the recipient's country. It's thus possible to define one fee for transfers within Denmark and another fee for a transfer from Denmark to the USA. A fee may be defined using the following data fields:
- minfee, Minimum fee charged for the operation
- basefee, Base fee charged for the operation
- share, Percentage of the amount involved with the operation that the end-user will be charged
The actual fee charged is defined as as BASEFEE + AMOUNT * SHARE or MINFEE, whichever is higher.
Additionally trigger: Modify_EndUserTxn_Proc has been changed to take fees into account when calculating an End-User's current balance.

Extended the data being logged to improve auditing to include:
- IP
- Mobile Number or E-Mail Address
IP has been added to both database table: Log.Transaction_Tbl and EndUser.Transaction_Tbl where as Mobile Number or E-Mail Address has only been added to database table: EndUser.Transaction_Tbl.
For EndUser.Transaction_Tbl.
This has affected the SQL Query in the following methods:
- General::newTransaction
- EndUserAccount::associate
- EndUserAccount::purchase
- EndUserAccount::topup
- Transfer::makeTransfer

git-svn-id: svn://svn.cellpointmobile.com/mpoint/server/trunk@91 86c063f6-31ef-4d33-b9ba-d2d3e4924f03