Fixed bug which caused the Order ID to be displayed for both SMS and E-Mail Receipts. This has affected the following methods: - Callback::sendSMSReceipt - EMailReceipt::sendReceipt
Fixed bugs in the SQL queries used to link transactions to end-users which caused the queries to throw an SQL Error.
These fixes have affected the SQL queries of the following methods from class: EndUserAccount:
- associate
- purchase
- topup

Fixed the following bugs in the back-end controller for transfers: sys/make_transfer.php.
- Invalid calculation of amount when fees where involved
- Reference to the wrong country configuration prevented Confirmations Codes from being sent via SMS
- Wrong calculation of the fee caused fees related to transfers to be extremely low

Fixed bug in the SQL query of method: getTxnHistory from class: Home which caused the fee to be returned as a negative number (which is how it's stored in database table: EndUser.Transaction_Tbl).

Fixed bug in method: initCallback from class: DIBS which caused the back-end controller for DIBS callbacks to fail for account top-ups using stored credit cards due to "clientid" not being part of the data sent to: callback/dibs.php via HTTP POST. initCallback now includes the correct clientid as part of the posted data.

Made the configuration for the default UA Profile URL dynamic in conf/global.php. Prior to setting the constant: sDEFAULT_UA_PROFILE a call to method: getBrowserType from class: General is made.
If getBrowserType returns "web" the constant is set to mPoint's default browser: http://iemendo.cellpointmobile.com/data/mpoint-ajax.xml.
This URL has also been registered in iEmendo as a valid browser.

git-svn-id: svn://svn.cellpointmobile.com/mpoint/server/trunk@92 86c063f6-31ef-4d33-b9ba-d2d3e4924f03