Added new automatic popup upon login which notifies the user if any of the following pieces of information are missing: - Mobile Number - E-Mail Address - Profile Info (Firstname / Lastname) The user is advised to go to "My Account" and add the missing information. Additionally this behaviour has been removed from the back-end controller for authentication: login/sys/auth.php
Changed all AJAX viewers to use the HTML "button" tag rather than the HTML "input" for all forms.
This has affected the following AJAX viewers:
- home/accept.xsl
- home/code.xsl
- home/confirm.xsl
- home/edit_email.xsl
- home/edit_mobile.xsl
- home/edit_password.xsl
- home/transfer.xsl
- login/otp.xsl
- new/step1.xsl
- new/step2.xsl

Fixed bug in the back-end controller for the Transfer component: home/sys/make_transfer.php which caused the unformatted amounts to be displayed for the error messages when min transfer or account balance was exceeded.

Fixed bug in method: convert from class: Transfer which was introduced when the method was changed to use method: getExchangeRates from class: Transfer.
Method: getExchangeRates is intended to normalise the XML Document with the exchange rates making it simple to add another / additional data source.

git-svn-id: svn://svn.cellpointmobile.com/mpoint/server/trunk@96 86c063f6-31ef-4d33-b9ba-d2d3e4924f03