Implemented HTML5 template for mPoint's payment flow. The new template uses the iScroll javascript library to emulate the behaviour of the CSS "overflow" style in order to enable localized scrolling on a page. The HTML5 template is designed to allow mPoint to be embedded into an advanced HTML5 based mobile web application which simulates a native application with a top bar and bottom bar (tab bar in iPhone terminology) that should always be shown. mPoint may be embedded into this type of application by using an iframe tag within the content area. The DIBS payment pages have further been updated to include an HTML5 version which slso uses the iScroll library to enable localized scrolling.
The addition of an HTML5 template have also meant that a new parameter must be passed to the static method: getMarkupLanguage from class: General in order to allow this method to determine whether the end-user's mobile phone supports HTML5.
The use of the HTML5 template may be enabled by a new setting: markup in mPoint's Client Configuration.

Improved the logic of mPoint's Mobile Web API so that an end-user which has one or more cards stored for a given client is automatically taken to the "My Account" page (cpm/payment.php) thereby saving a click.
This change has affected all .php files in mPoint which may be accessed from a mobile phone.
This has also meant that a "New Card" button has been added to the "My Account" page allowing the end-user to access the list of available payment options.

git-svn-id: svn://svn.cellpointmobile.com/mpoint/server/trunk@103 86c063f6-31ef-4d33-b9ba-d2d3e4924f03