Improved the error logging in method: refund from class: DIBS to log the entire response from DIBS if the refund operations fails.
Added support to method: valBasic from class: Validate for addressing a bug in iOS5 which will cause a large integer to be sent as a negative number.

Created preliminary implementation of an API directed at a native SDK for mPoint. The intention of the new API is to support a completely native user experience for Mobile Apps. and provide a mobile optimized user experience for 3D Secure.
The new API is intended to be called from the mPoint SDK.

Added documentation for new Payment Service Provider: Wannafind

Added 3D Secure documentation for WorldPay

git-svn-id: svn://svn.cellpointmobile.com/mpoint/server/trunk@107 86c063f6-31ef-4d33-b9ba-d2d3e4924f03