Implemented support for Oracle, this has meant that all SQL queries have been changed to use enabled = '1' rather than enabled = true. This change has affected to following classes: - Admin - Capture - ClientConfig - CountryConfig - CreditCard - EndUserAccount - General - Home - MyAccount - Products - PSPConfig - Refund - Shipping - SMSPurchase - SurePay - TopUp - Transfer - TxnInfo - Validate
Added mapping for ECMC_DEBIT-SSL to MasterCard (Card ID = 7) and VISA_DEBIT-SSL to VISA (Card ID = 8) in class: WorldPay.

Added the following new APIs to the mApp module:
- mApp/api/find_address.php
- mApp/api/get_transaction-history.php
- mApp/api/login.php
- mApp/api/save_address.php
- mApp/api/save_info.php
- mApp/api/transfer.php
- mApp/api/verify.php

Implemented a new concept: URL Hooks, which enables mPoint to hook into external systems through the Mobile Enterprise Service Bus. Currently this is used by the "Save Account" API to import data from an external system.
This has added a new class: ClientInfo, which is used to represent the "client-info" for all requests made to APIs that are part of the mApp module.


git-svn-id: svn://svn.cellpointmobile.com/mpoint/server/trunk@222 86c063f6-31ef-4d33-b9ba-d2d3e4924f03