Fixed bug when finding the user's account id during "login" which would result in the user's payment card not being returned because method: getAccountID from class: Home would simply take the first end-user account without taking into consideration whether the account was accessible for the merchant. This has added a new optional parameter: $clid to method: getAccountID from class: Home.
Fixed bug in the "Initialize Payment" API, which had markup hardcoded as "app" when creating the transaction.
The transaction is now created with the Account configuration (which should be "app" or "html5" when using this API).

Implemented support for using PayEx as the Payment Service Provider in the "Pay" API.

Changed all amounts for the "Transfer" API to be in the country's smallest currency (øre, cents, pence etc.) rather than the country's largest (DKK, USD, GBP etc.). Additionally the API has been extended to accept a message.

Added the following new APIs:
- Get Account Info: /mApp/api/get_account.php
- Delete Card: /mApp/api/del_card.php
This has also required updates to the XML schema with the protocol definitions: mpoint.xsd

Fixed bug in method: savePassword from class: EndUserAccount, which would use the client's country configuration rather than the user's when finding the user's account.

Extended the WorldPay integration to also forward the customer's email address if available. This enables WorldPay to use the email address as part of their risk scoring that is used to identify fraud.

git-svn-id: svn://svn.cellpointmobile.com/mpoint/server/trunk@270 86c063f6-31ef-4d33-b9ba-d2d3e4924f03