Fixed bug in the signature for method: sendNewAccountSMS in class: Transfer, which would cause a crash.
Extended method: saveEMail from class: EndUserAccount so it accepts a CountryConfig instance to support saving email addresses for users that are not in the same country as the client has been configured.

Updated the Modify_EndUserTxn_Proc trigger so it filters out cancelled transfers when calculating the balance on the end-user's account.

Fixed bug in the controller for the "Get Transaction History" mApp API, which would return the wrong currency if the user was from a different country than the client configuration.

Implemented support for Accept URL in the controller for the "Initialize Payment" mAPP API. The value is currently taken from the client configuration.

Implemented support for saving an e-mail address when creating an account through the "Save Account" mApp API.

git-svn-id: svn://svn.cellpointmobile.com/mpoint/server/trunk@316 86c063f6-31ef-4d33-b9ba-d2d3e4924f03