Implemented support for storing cards through WorldPay. This has added a new field: stored_card to database table: Client.MerchantAccount_Tbl, which identifies the merchant account as being dedicated to processing payments using a stored card or for card details. This has also added a new step to the payment flow represented by the pay/store.php controller which prompts the customer to store their card details
Implemented support for Single Sign-On by invoking an external web service to authenticate the customer.
Single Sign-On may be enabled by supplying an auth-token when initializing a payment request and by either supplying an auth-url or having the URL configured in mPoint.

Updated the API documentation to describe the new Single Sign-On capabilities.

git-svn-id: svn://svn.cellpointmobile.com/mpoint/server/trunk@343 86c063f6-31ef-4d33-b9ba-d2d3e4924f03