Implemented support for handling Card Holder Name for PayEx in method: initialize from class: PayEx.
Fixed bug in method: complete from class: PayEx, which would log a rejected authorization as successful.

Fixed bug in method: produceInfo from class: TxnInfo, which wouldn't transfer the configured Authentication URL from the Client Configuration to the Transaction Information.

Fixed bug in the controller: /pay/sys/save_name.php, which would cause a PHP error to be thrown the first time a card was saved using the Customer Ref to identify the account.

git-svn-id: svn://svn.cellpointmobile.com/mpoint/server/trunk@366 86c063f6-31ef-4d33-b9ba-d2d3e4924f03