Performed code clean-up of all methods in class: Admin.
Renamed method: getTxnIDFromExtId to: getTxnIDFromExtID in class: Callback.
Additionally the SQL query now correctly escapes the external ID so it cannot be used for SQL injection attacks.
This has also affected the Callback controller for NetAxept: callback/netaxept.php

Performed code clean-up of method: authTicket in class: CPG.
Additionally numerous bugs, which caused the Order XML request to be rejected by CPG have been fixed.
This has added a new private method: _getCountryCode, which translates from mPoint's Country ID to Emirates' Country Code.

Implemented support for "automagically" setting the timestamp format for Oracle in the constructor in class: General.

Implemented workaround for a bug in PHP's unserialize function in method: getMessageData from class: General.
The PHP bug would prevent unserialization for arrays when values contained one of the following characters: ', " or ;

Implemented compatibility with Oracle (and other databases) in method: getAccountInfo from class: Home.
This has removed the usage of the PostGreSQL method: Extract from the SQL query.

Fixed bug in method: produceInfo from class: TxnInfo, which would result in the Description and IP being empty.

Fixed bug in method: authTicket from class: WorldPay, which would result in mPoint not recognizing the Payment when receiving a callback notification from WorldPay if the callback was received before the "Authorized" response.
This fix means that mPoint no longer must generate an internal callback in the controller: /cpm/sys/pay_account.php when paying with a Stored Card but rather should rely on the Asyncronous callback from WorldPay.

Fixed bug in the API controller for mPoint's "Authorize Payment" API in the mApp module, which would result in an error when authorizing a Stored Card through CPG.

Fixed bugs in the Audit Log implementation in the following API controllers from mPoint's mApp module:
- /mApp/api/login.php
- /mApp/api/del_card.php
- /mApp/api/save_card.php

Fixed bug in the API controller for mPoint's "Initialize Payment" API in the mApp module, which would result in Custom Variables not being correctly stored. Additionally a bug in the mPoint's protocol definition: mpoint.xsd has been fixed.

Implemented support for automatically redirecting the customer to the merchant's site in the "Payment Accepted" viewer.

git-svn-id: svn://svn.cellpointmobile.com/mpoint/server/trunk@546 86c063f6-31ef-4d33-b9ba-d2d3e4924f03