Fixed bugs in method: authTicket from class: CPG, which would pass the wrong currency code in the Order XML.
Added trimming of the card number mask to the following methods from class: EndUserAccount:
- saveCard
- getCardIDFromCardDetails

Fixed bug in the SQL query in method: logTransaction from class: General, which would use the Country Configuration from the Client Configuration rather than from the Transaction Info. This would result in the wrong country being logged for the transaction, which in turn would cause problems when determining the currency.

Fixed bugs in methods: produceInfo and toXML from class: TxnInfo, which would use the Country Configuration from the Client Configuration rather than from the Transaction Info. This would result in the wrong country being logged for the transaction, which in turn would cause problems when determining the currency.

Fixed bugs in the API controller for mPoint's "Initialize Payment" API in the mApp module, which would take the wrong country configuration when initializing a new payment transaction. This would result in the wrong country being logged for the transaction, which in turn would cause problems when determining the currency.

Implemented more graceful handling of a 4 digit expiry year in the API controller for mPoint's "Save Card" API in the mApp module.

Cleaned up the request format in the API controller for mPoint's "Search" API in the mConsole module.
Additionally bugs have been fixed when constructing the SQL query in method: getAuditLog from class: General, which would construct a start / end date, which wasn't compatible with Oracle.

git-svn-id: svn://svn.cellpointmobile.com/mpoint/server/trunk@551 86c063f6-31ef-4d33-b9ba-d2d3e4924f03